\name{RcppArmadillo-Functions}
\alias{cpp1}
\alias{cpp2}
\alias{cpp3}
\alias{cpp4}
\alias{cpp5}
\alias{cpp6}
\alias{cpp7}
\alias{cpp8}
\title{Set of functions in example RcppArmadillo package}
\description{
  These four functions are created when
  \code{RcppArmadillo.package.skeleton()} is invoked to create a
  skeleton packages.
}
\usage{
cpp1(x)
cpp2(x)
cpp3(x)
cpp4(x)
cpp5(x)
cpp6(x)
cpp7(x)
cpp8(x)
}
\arguments{
  \item{x}{a numeric vector}
}
\value{
  \code{cpp1()} is cpp function of \code{x}.

  \code{cpp2()} is cpp function of \code{x}.
  
  \code{cpp3()} is cpp function of \code{x}.
  
  \code{cpp4()} is cpp function of \code{x}.
  
  \code{cpp5()} is cpp function of \code{x}.
  
  \code{cpp6()} is cpp function of \code{x}.
  
  \code{cpp7()} is cpp function of \code{x}.
  
  \code{cpp8()} is cpp function of \code{x}.
 
}
\details{
  SV model
}
\references{
  See the documentation for Armadillo, and RcppArmadillo, for more details.
}
\author{Peng Sun}

